+++ src/fs.spec.js
@@ -1,5 +1,5 @@
 import '@babel/polyfill'; // Required for NodeJS < 10
-import { lStat, isDir, readDir } from './fs';
+import { lStat, isDir, readDir, stat } from './fs';
 import {
   file1a,
   level2,
@@ -8,6 +8,69 @@ import {
   linkToUnexistingFile,
   unexistingFile,
 } from '../fixture';
+
+/** @test {stat} */
+describe('stat', () => {
+  it('is a function', () => {
+    expect(stat).toBeFunction();
+  });
+
+  it('returns a Promise', () => {
+    expect(stat(file1a)).toBeInstanceOf(Promise);
+  });
+
+  it('resolves when file exists', async () => {
+    await expect(stat(file1a)).toResolve();
+  });
+
+  it('resolves to the correct stat when file exists', async () => {
+    const stats = await stat(file1a);
+    expect(stats.isFile()).toBe(true);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(false);
+  });
+
+  it('resolves when directory exists', async () => {
+    await expect(stat(level2)).toResolve();
+  });
+
+  it('resolves to the correct stat when directory exists', async () => {
+    const stats = await stat(level2);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(true);
+    expect(stats.isSymbolicLink()).toBe(false);
+  });
+
+  it('rejects when file does not exist', async () => {
+    await expect(stat(unexistingFile)).toReject();
+  });
+
+  it('resolves when symlink links to an existing file', async () => {
+    await expect(stat(linkToSiblingFile)).toResolve();
+  });
+
+  it('resolves to the correct stat when symlink links to an existing file', async () => {
+    const stats = await stat(linkToSiblingFile);
+    expect(stats.isFile()).toBe(true);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(false);
+  });
+
+  it('resolves when symlink links to an existing directory', async () => {
+    await expect(stat(linkToSiblingDirectory)).toResolve();
+  });
+
+  it('resolves to the correct stat when symlink links to an existing directory', async () => {
+    const stats = await stat(linkToSiblingDirectory);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(true);
+    expect(stats.isSymbolicLink()).toBe(false);
+  });
+
+  it('rejects when symlink links to unexisting file', async () => {
+    await expect(stat(linkToUnexistingFile)).toReject();
+  });
+});
 
 /** @test {lStat} */
 describe('lStat', () => {
@@ -24,10 +87,10 @@ describe('lStat', () => {
   });
 
   it('resolves to the correct stat when file exists', async () => {
-    const stat = await lStat(file1a);
-    expect(stat.isFile()).toBe(true);
-    expect(stat.isDirectory()).toBe(false);
-    expect(stat.isSymbolicLink()).toBe(false);
+    const stats = await lStat(file1a);
+    expect(stats.isFile()).toBe(true);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(false);
   });
 
   it('resolves when directory exists', async () => {
@@ -35,10 +98,10 @@ describe('lStat', () => {
   });
 
   it('resolves to the correct stat when directory exists', async () => {
-    const stat = await lStat(level2);
-    expect(stat.isFile()).toBe(false);
-    expect(stat.isDirectory()).toBe(true);
-    expect(stat.isSymbolicLink()).toBe(false);
+    const stats = await lStat(level2);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(true);
+    expect(stats.isSymbolicLink()).toBe(false);
   });
 
   it('rejects when file does not exist', async () => {
@@ -50,10 +113,10 @@ describe('lStat', () => {
   });
 
   it('resolves to the correct stat when symlink links to an existing file', async () => {
-    const stat = await lStat(linkToSiblingFile);
-    expect(stat.isFile()).toBe(false);
-    expect(stat.isDirectory()).toBe(false);
-    expect(stat.isSymbolicLink()).toBe(true);
+    const stats = await lStat(linkToSiblingFile);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(true);
   });
 
   it('resolves when symlink links to an existing directory', async () => {
@@ -61,21 +124,21 @@ describe('lStat', () => {
   });
 
   it('resolves to the correct stat when symlink links to an existing directory', async () => {
-    const stat = await lStat(linkToSiblingDirectory);
-    expect(stat.isFile()).toBe(false);
-    expect(stat.isDirectory()).toBe(false);
-    expect(stat.isSymbolicLink()).toBe(true);
+    const stats = await lStat(linkToSiblingDirectory);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(true);
   });
 
   it('resolves when symlink links to unexisting file', async () => {
     await expect(lStat(linkToUnexistingFile)).toResolve();
   });
 
-  it('resolves to the correct stat symlink links to unexisting file', async () => {
-    const stat = await lStat(linkToUnexistingFile);
-    expect(stat.isFile()).toBe(false);
-    expect(stat.isDirectory()).toBe(false);
-    expect(stat.isSymbolicLink()).toBe(true);
+  it('resolves to the correct stat when symlink links to unexisting file', async () => {
+    const stats = await lStat(linkToUnexistingFile);
+    expect(stats.isFile()).toBe(false);
+    expect(stats.isDirectory()).toBe(false);
+    expect(stats.isSymbolicLink()).toBe(true);
   });
 });
 
@@ -120,15 +183,6 @@ describe('isDir', () => {
 
 /** @test {readDir} */
 describe('readDir', () => {
-  const expectedFiles = [
-    'file2a',
-    'file2b',
-    'level3',
-    'link-to-directory',
-    'link-to-grand-parent-directory',
-    'link-to-parent-directory',
-  ];
-
   it('is a function', () => {
     expect(readDir).toBeFunction();
   });
@@ -142,7 +196,7 @@ describe('readDir', () => {
   });
 
   it('resolves to the list of files when directory exists', async () => {
-    expect(await readDir(level2)).toEqual(expectedFiles);
+    expect(await readDir(level2)).toMatchSnapshot();
   });
 
   it('rejects when file does not exist', async () => {
@@ -154,7 +208,7 @@ describe('readDir', () => {
   });
 
   it('resolves to the list of files when symlink links to an existing directory', async () => {
-    expect(await readDir(linkToSiblingDirectory)).toEqual(expectedFiles);
+    expect(await readDir(linkToSiblingDirectory)).toMatchSnapshot();
   });
 
   it('rejects when symlink links to unexisting file', async () => {