+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. The format 
 
 - `directoriesFirst` option - [#8]
 - `showDirectories` option - [#9]
+- `followSymlinks` option - [#11]
 
 ## [v0.2.0] - 2018-11-23
 
@@ -50,6 +51,7 @@ All notable changes to this project will be documented in this file. The format 
 [#7]: https://github.com/amercier/files-by-directory/pull/7
 [#8]: https://github.com/amercier/files-by-directory/pull/8
 [#9]: https://github.com/amercier/files-by-directory/pull/9
+[#11]: https://github.com/amercier/files-by-directory/pull/11
 [v0.1.0]: https://github.com/amercier/files-by-directory/compare/v0.0.0...v0.1.0
 [v0.1.1]: https://github.com/amercier/files-by-directory/compare/v0.1.0...v0.1.1
 [v0.1.2]: https://github.com/amercier/files-by-directory/compare/v0.1.1...v0.1.2